Organize other client methods
