+++ platiagro/__init__.py
@@ -8,8 +8,6 @@ from platiagro.figures import list_figures, save_figure
 from platiagro.metrics import list_metrics, save_metrics
 from platiagro.models import load_model, save_model
 from platiagro.tasks import create_task
-from platiagro.deployments import list_projects, get_project_by_name, list_deployments, \
-    get_deployment_by_name, run_deployments
 
 __all__ = ["download_artifact",
            "list_datasets", "load_dataset", "save_dataset", "stat_dataset",
@@ -19,7 +17,6 @@ __all__ = ["download_artifact",
            "list_figures", "save_figure",
            "list_metrics", "save_metrics",
            "load_model", "save_model",
-           "create_task",
-           "list_projects", "get_project_by_name", "list_deployments", "get_deployment_by_name", "run_deployments"]
+           "create_task"]
 
 __version__ = "0.2.0"
